Fix desktop file names for eos-app- application overrides
In order for desktop matching to work correctly, we need to make sure
desktop file overrides we install have the same last component as the
original desktop files installed by the application packages.
This way, we can just use eos-app- as a vendor prefix, and use the
regular application matching logic already present in the shell.

[endlessm/eos-shell#242]