Revert GCompris launching to full screen
There is a glitch in GCompris window mode that apparently closes the first
window and opens a second one.  This glitch triggers the shell to show
the app selector, and thus GCompris effectively launches minimized.

This will need further investigation, and perhaps a change to
the source code for GCompris.  For now, we are better off launching
in the default full screen mode.

[endlessm/eos-shell#497]