Replace epiphany with eos-browser
Replace browser application displayed on the desktop.
Change script that generates desktop entries for links.

Note the use of "eos-browser" rather than just "browser"
in the applications.csv file, so that the resulting
"eos-app-eos-browser.desktop" file follows the rule of vendor prefix
followed by original desktop entry name, which is important for the
logic used with mini icons on the taskbar.

[endlessm/eos-shell#639]