Auto import user's media
For Shotwell, the upstream default of auto-import is false,
so this is necessary to enable that feature by default.

For Rhythmbox, the upstream default of monitor-library is true,
but we explicitly set it here for consistency
(and just in case the upstream default were to change).

In either case, we don't explicitly set the import location,
as the upstream defaults properly point to the user's home folders,
and any attempt to explicitly set the folder name would
not properly handle translations of the home folder names.

[endlessm/eos-shell#849]