Disable splash screen for some applications
Adds a column to the applications.csv names SplashScreen.
Currently, the supported values are a blank field for the
default splash screen, or the special value 'none' indicating
that no splash screen is displayed when launching the app.

For now, all applications have the default splash screen
except for the calculator and skype.

[endlessm/eos-shell#895]