Apps that request no title bar when maximized now retain the bar when tiled
The theme was already set to ignore requests by apps to hide the
title bar when they are maximized. However, when apps chose to
hide the title bar when maximized, the title bar would in fact be
hidden when the app was tiled to the left or right.

This patch also fixes a problem where the style looked slightly
weird on some apps that were maximized but not focused, which
appears to be due to a copy/paste error in the theme.

[endlessm/eos-shell#989]