Remove redundant list of background images
The per-personality background image lists are generated
in eos-media from the available images.
There is no reason to try to maintain such a list in eos-theme.

Furthermore, since some personalities may have a copy of
the default background image with a different name,
in order to avoid duplicates in the selection list,
rename endlessos.xml to endlessos-backgrounds-default.xml,
so that it is only available to the default personality.

[endlessm/eos-shell#1584]