build: Workaround issue with automake 1.15+
In automake 1.15+, a 2 level build directory of _build/sub is used for
distcheck. This is to try to flush out issues with programs making the
wrong assumptions about srcdir != builddir. Unfortunately, one of these
programs is intltool:

  The following files contain translations and are currently not in use.
  Please consider adding these to the POTFILES.in file, located in the
  po/ directory.

  sub/icons/index.theme.in

  If some of these files are left out on purpose then please add them to
  POTFILES.skip instead of POTFILES.in.

Workaround this by adding this path to POTFILES.skip even though it's
entirely specific to distcheck. No fix has been committed upstream, and
this is how other projects are managing the issue.

https://bugs.launchpad.net/intltool/+bug/1117944
https://phabricator.endlessm.com/T18349