add new --enable-app-icons, default to off
We only want to ship our app override icons in the Flatpak theme
rather than in the OS, so we split the app icons (currently just
GNOME Weather) into a separate folder and configure flag which
defaults to off. The OS will not ship them, but we'll turn it
on in the Flatpak extension build.

https://phabricator.endlessm.com/T20599