settings: Add read-only dconf defaults database via OSTree
We currently have a read-only dconf defaults database in
/var/lib/eos-image-defaults, but that can only be set by the image
builder (due to being in /var). In order to ship changing defaults in
the OSTree, we need an additional dconf defaults database in
/usr/share/eos-image-defaults. This commit adds that.

The database is, initially, empty. A following commit will add the first
key which this feature is needed for.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20990