settings: Add default dconf value for power notifications
Prevent them being shown on the lock screen, since it’s pointless to
tell the user that their computer is about to suspend if they’re not
using it. Furthermore, the notification (by default) doesn’t show any
details – just ‘! Power’ – which adds to the unnecessary mystery. Best
just to hide it.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20990