settings: set touchpad click-method=default
Upstream GNOME has gone to click-method=fingers and after trying for a month in
Endless, we have concluded this makes anyone with prior Windows or PC
experience conclude their touchpad is broken, and try to return their computer,
post on the forum, stop using Endless, fail OEM tests, etc. We think it's the
wrong call considering Endless is unlikely to enter the same channels /
geographies as Macs, Chromebooks or super high-end Ultrabooks.

https://phabricator.endlessm.com/T25142