safe-defaults: introduce new mechanism with OpenDNS family shield
Introduce a new safe-defaults subdirectory to add a group of config files in
the OSTree which can be symlinked into /etc to activate the family defaults. An
eos-safe-defaults command with enable/disable arguments uses systemd-tmpfiles
to create the symlinks from /etc to /usr/share to create or remove all of the
links at once. The systemd job is provided to be run as an upgrade unit,
designed to only be enabled on systems which are expected to always have
safe defaults enabled, but the eos-safe-defaults (enable|disable) command
can be used on any system.

The first configuration fragment is to use NetworkManager's Global DNS
configuration to set the system-wide DNS servers to OpenDNS Family Shield.

https://phabricator.endlessm.com/T23464