settings: re-enable Nautilus search provider by default
This was disabled by default in
2a21c39dc2a7598c589337c603049eb85281e248. The rationale given was:

> We want our global search to feel more like google where you search
> for things like "Mexico", "rabbit" and what not, and less like a
> traditional system search where you look for "my-list.txt".

I think I understand the sentiment here, but I think the cost (making it
harder to discover your files) is too high. The lack of local file
search was reported as a bug/feature request by two people within
Endless, and I think that as our target audience has moved further from
“inexperienced, mostly-offline computer user” the need for this feature
has increased somewhat.

Remove it from the 'disabled' list, but leave it naturally sorted within
the other search provides not listed in the sort-order. Users can
reorder or disable search providers in Settings, but it's easier to
disable a feature you don't like than it is to enable a feature you
didn't know you had.

https://phabricator.endlessm.com/T28224