autostart: Add autostart desktop file to set EKN flatpak permissions
Flatpak (correctly) warns the user about EKN multiplexer continuing to
run as a background process. As the multiplexer is designed to do that,
set the permissions in the xdg-desktop-portal database so that the
multiplexer is allowed to run in the background, and the user isn’t
notified about it.

It has to be done on startup so all users have the permissions set in
their session (xdg-desktop-portal provides no system-wide configuration
hook), and it can’t be done from within Hack as that would require
breaking the flatpak sandboxing.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T29677