gnome-software: Install a background image for the upgrade banner
We no longer ship this as a downstream customisation within the
gnome-software source tree. Instead, since support for it has been added
upstream, we can ship the background image separately and it will be
loaded by gnome-software at runtime.

This background image is used for the banner which is shown in
gnome-software when an OS update is available.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T33015