Revert "Decrease the likelihood that an unsteady click is interpreted as a drag"
This reverts commit ffcedd7efff2ab7a14e62849532c8ab8421c2548.

Similar to #388, the [original request] was from over a decade ago and
looks to be working around a very specific hardware issue combined with
the design of EOS at the time. Hardware has improved, it's possible drag
detection has improved, and the app grid is completely different these
days as well.

Let's inherit the upstream defaults here; if we discover a usability
issue from the field, we should bring that feedback upstream to GNOME
since we're involved with the design team there.

[original request]: https://github.com/endlessm/eos-shell/issues/517