Install systemwide gnome-initial-setup vendor.conf file
This file is moved from eos-image-builder. The copy there also
contained:

    [pages]
    run_welcome_tour=false

but this key has not been read by Initial Setup for many years.

Rather than shipping this default list of languages as a mutable
customisation in every image, we will now ship the defaults in the
immutable OS and only override it if necessary for a particular image.

https://phabricator.endlessm.com/T35040